+++ apps/react-demo-app/src/utils/sidecar.ts
@@ -82,7 +82,10 @@ const generateBaseUrls = (cluster: string, prod = false) => {
     projectId: '1ee63b21-27c7-44ad-891f-4bd9af378b72', // <- move this to release-configs
     version: 'Production', // <- move this to release-configs
   },
-  intercom: 'ou1uyk2p',
+  intercomSettings: {
+    app_id: 'ou1uyk2p',
+    hide_default_launcher: true,
+  },
   ...((window as any).__cogniteSidecar || {}),
 };