+++ packages/intercom-helper/src/methods/identityVerification.spec.ts
@@ -18,15 +18,22 @@ describe('get Hmac key from server', () => {
     };
   });
 
-  it('identityVerification resolves', () => {
+  it('identityVerification resolves with no project', () => {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     (getHmac as any).mockImplementation(() =>
       Promise.resolve({ hmac: 'hmacHash1234', userUid: 'Unique ID' })
     );
     return identityVerification({
-      appsApiUrl: 'test',
+      appsApiUrl: 'fakeUrl',
       headers: { Authorization: 'test' },
     }).then((response) => {
+      expect(getHmac).toHaveBeenCalledWith(
+        'fakeUrl',
+        {
+          Authorization: 'test',
+        },
+        undefined
+      );
       expect(response).toEqual({ success: true });
       expect(global.window.Intercom).toHaveBeenCalledWith('update', {
         user_id: 'Unique ID',
@@ -35,14 +42,67 @@ describe('get Hmac key from server', () => {
     });
   });
 
-  it('identityVerification rejects', () => {
+  it('identityVerification rejects with no project', () => {
     const error = new Error('failure');
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     (getHmac as any).mockImplementation(() => Promise.reject(error));
     return identityVerification({
-      appsApiUrl: 'test',
+      appsApiUrl: 'fakeUrl',
       headers: { Authorization: 'test' },
     }).then((response) => {
+      expect(getHmac).toHaveBeenCalledWith(
+        'fakeUrl',
+        {
+          Authorization: 'test',
+        },
+        undefined
+      );
+      expect(response).toEqual({ success: false, error });
+      expect(global.window.Intercom).not.toHaveBeenCalled();
+    });
+  });
+
+  it('identityVerification resolves with project', () => {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    (getHmac as any).mockImplementation(() =>
+      Promise.resolve({ hmac: 'hmacHash1234', userUid: 'Unique ID' })
+    );
+    return identityVerification({
+      appsApiUrl: 'fakeUrl',
+      headers: { Authorization: 'test' },
+      project: 'myFakeProject',
+    }).then((response) => {
+      expect(getHmac).toHaveBeenCalledWith(
+        'fakeUrl',
+        {
+          Authorization: 'test',
+        },
+        'myFakeProject'
+      );
+      expect(response).toEqual({ success: true });
+      expect(global.window.Intercom).toHaveBeenCalledWith('update', {
+        user_id: 'Unique ID',
+        user_hash: 'hmacHash1234',
+      });
+    });
+  });
+
+  it('identityVerification rejects with project', () => {
+    const error = new Error('failure');
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    (getHmac as any).mockImplementation(() => Promise.reject(error));
+    return identityVerification({
+      appsApiUrl: 'fakeUrl',
+      headers: { Authorization: 'test' },
+      project: 'myFakeProject',
+    }).then((response) => {
+      expect(getHmac).toHaveBeenCalledWith(
+        'fakeUrl',
+        {
+          Authorization: 'test',
+        },
+        'myFakeProject'
+      );
       expect(response).toEqual({ success: false, error });
       expect(global.window.Intercom).not.toHaveBeenCalled();
     });